+++ internal/route/sync_other.go
@@ -6,5 +6,5 @@ import "fmt"
 
 // SyncRoutes is not supported outside Windows.
 func SyncRoutes(connectionName string, desired []string) error {
-	return fmt.Errorf("sinkronisasi rute VPN hanya didukung di Windows")
+	return fmt.Errorf("VPN route sync is only supported on Windows")
 }